+++ examples/rpi-serial.rs
@@ -1,13 +1,13 @@
 use linux_embedded_hal::Serial;
 use std::{path, thread, time};
 
-use jrk_g2_rs::{JrkG2, JrkG2Serial};
+use jrk_g2_rs::{JrkG2, Serial as Jrk};
 
-fn main() {
-    let serial = Serial::open(path::Path::new("/dev/ttyS0")).unwrap();
-    let mut jrk = JrkG2Serial::new(serial);
+fn main() -> Result<(), linux_embedded_hal::serial_core::Error> {
+    let serial = Serial::open(path::Path::new("/dev/ttyS0"))?;
+    let mut jrk = Jrk::new(serial);
     println!("jrk initialized on rpi by serial");
-    let mut ret: String = "".to_string();
+    let mut ret: String = String::new();
 
     loop {
         if let Err(e) = jrk.stop_motor() {
@@ -16,7 +16,7 @@ fn main() {
         thread::sleep(time::Duration::from_secs(2));
         jrk.show_vars(&mut ret).ok();
         println!("{ret}");
-        ret = "".to_string();
+        ret = String::new();
 
         if let Err(e) = jrk.set_target(1450) {
             println!("SerialError: {e:?}");
@@ -24,6 +24,6 @@ fn main() {
         thread::sleep(time::Duration::from_secs(2));
         jrk.show_vars(&mut ret).ok();
         println!("{ret}");
-        ret = "".to_string();
+        ret = String::new();
     }
 }