+++ src/blocking_i2c.rs
@@ -4,16 +4,16 @@ use embedded_hal::blocking::i2c;
 use nb::block;
 
 /// Implement the `JrkG2` trait for Blocking I2C
-pub struct JrkG2BlockingI2c<I2c> {
+pub struct BlockingI2c<Bus> {
     device: u8,
-    i2c: I2c,
+    i2c: Bus,
 }
 
-impl<I2c, I2cError> JrkG2BlockingI2c<I2c>
+impl<Bus, I2cError> BlockingI2c<Bus>
 where
-    I2c: i2c::Write<Error = nb::Error<I2cError>> + i2c::Read<Error = nb::Error<I2cError>>,
+    Bus: i2c::Write<Error = nb::Error<I2cError>> + i2c::Read<Error = nb::Error<I2cError>>,
 {
-    pub fn new(i2c: I2c) -> Self {
+    pub const fn new(i2c: Bus) -> Self {
         Self { device: 0x0B, i2c }
     }
     /// The controller have a default 0x0B I2C address, but this can be manually changed in the
@@ -23,9 +23,9 @@ where
     }
 }
 
-impl<I2c, I2cError> JrkG2<I2cError> for JrkG2BlockingI2c<I2c>
+impl<Bus, I2cError> JrkG2<I2cError> for BlockingI2c<Bus>
 where
-    I2c: i2c::Write<Error = nb::Error<I2cError>> + i2c::Read<Error = nb::Error<I2cError>>,
+    Bus: i2c::Write<Error = nb::Error<I2cError>> + i2c::Read<Error = nb::Error<I2cError>>,
 {
     fn write(&mut self, data: &[u8]) -> Result<(), I2cError> {
         block!(self.i2c.write(self.device, data))