+++ src/serial.rs
@@ -4,30 +4,30 @@ use embedded_hal::serial;
 use nb::block;
 
 /// Implement the `JrkG2` trait for Serial
-pub struct JrkG2Serial<Serial> {
-    serial: Serial,
+pub struct Serial<Bus> {
+    serial: Bus,
 }
 
-impl<Serial> JrkG2Serial<Serial>
+impl<Bus> Serial<Bus>
 where
-    Serial: serial::Write<u8> + serial::Read<u8>,
+    Bus: serial::Write<u8> + serial::Read<u8>,
 {
-    pub fn new(serial: Serial) -> Self {
+    pub const fn new(serial: Bus) -> Self {
         Self { serial }
     }
 }
 
-impl<Serial> JrkG2<<Serial as serial::Read<u8>>::Error> for JrkG2Serial<Serial>
+impl<Bus> JrkG2<<Bus as serial::Read<u8>>::Error> for Serial<Bus>
 where
-    Serial: serial::Read<u8> + serial::Write<u8>,
+    Bus: serial::Read<u8> + serial::Write<u8>,
 {
-    fn write(&mut self, data: &[u8]) -> Result<(), <Serial as serial::Read<u8>>::Error> {
+    fn write(&mut self, data: &[u8]) -> Result<(), <Bus as serial::Read<u8>>::Error> {
         for &b in data.iter() {
             block!(self.serial.write(b)).ok(); // Infallible
         }
         Ok(())
     }
-    fn read(&mut self, cmd: VarOffset) -> Result<u16, <Serial as serial::Read<u8>>::Error> {
+    fn read(&mut self, cmd: VarOffset) -> Result<u16, <Bus as serial::Read<u8>>::Error> {
         self.write(&[JrkG2Command::GetVariable16 as u8 | (cmd as u8 + 1)])?;
         let l = block!(self.serial.read())?;
         let h = block!(self.serial.read())?;